Updated managers switch
Updated the switch between game modes in the game manager.
Added comments to conditions and requirements.

+++ Assets/Code/Scripts/Condition.cs
@@ -1,7 +1,20 @@
 [System.Serializable]
 public class Condition
 {
-    public string name;
-    public int min;
-    public int max;
-}
+    public string name;         // Stat name
+    // Only one it's used per condition
+    public float min;           // Min stat value to be valid
+    public float max;           // Max stat value to be valid
+
+    /*
+     * Name values:
+     * alcohol-level
+     * happiness
+     * anger
+     * anxiety
+     * fear
+     * confident
+     * tenderness
+     * energy
+     */
+}
